% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgelist_functions.R
\name{append_feature_annotations}
\alias{append_feature_annotations}
\title{Merge a table of additional feature-data into the feature-data (`$genes`) of
a `DGEList`}
\usage{
append_feature_annotations(dge, annotations, feature_id = "feature_id",
  ...)
}
\value{
a DGEList
}
\description{
If the `$genes` entry is undefined, add the feature data as the genes entry
}
\details{
Ensures that the feature-id / gene-id matches up between the provided
data-frame and the existing `DGEList`
}
