% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_counts.R
\name{cbind_feature_counts}
\alias{cbind_feature_counts}
\title{Combine a list of feature-count `data.frame`s into a single feature-count
`data.frame`}
\usage{
cbind_feature_counts(count_list, names_as_colnames = FALSE)
}
\arguments{
\item{count_list}{A `list` of `data.frame`s. Each `data.frame`
should have columns `feature_id`, `length` and a final column that
contains the read counts for some sample.}

\item{names_as_colnames}{Boolean. If `TRUE`, the column IDs in the
output `data.frame` are taken from the `list` names in `count_list`.}
}
\description{
The input `data.frame`s should have columns including `feature_id`, `length`
and a set of file paths. The file path column contains the read count for a
given feature in that sample. Returns a `data.frame` with the count column
for each sample juxtaposed and with `feature_id` and `length` columns.
}
