% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeglm_functions.R
\name{compute_fitted_values}
\alias{compute_fitted_values}
\title{Obtain fitted values for a given combination of design coefficients}
\usage{
compute_fitted_values(x, contrasts, center = FALSE)
}
\arguments{
\item{x}{A fit from running an edgeR pipeline, eg, a
DGEGLM object.}

\item{contrasts}{A matrix defining linear combination of design
coefficients. eg, this might define the combination of design coefs that
gives the fitted value for a particular (set of) treatment group. This
should be specified in the format of a contrasts matrix (a row for each
design coefficient, a column for each computed value). If rownames are
specified, they should be identical to those used in the corresponding
design matrix.}

\item{center}{Boolean: should the mean value be subtracted from
each row before returning the results matrix?}
}
\description{
Obtain fitted values for a given combination of design coefficients
}
