% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt_io.R
\name{condense_comma_separated_numbers}
\alias{condense_comma_separated_numbers}
\title{Converts comma-containing strings into numbers}
\usage{
condense_comma_separated_numbers(x)
}
\arguments{
\item{x}{A vector of comma-containing strings, to be
converted to numbers.}
}
\description{
Converts comma-containing strings into numbers
}
