% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgelist_functions.R
\name{cqn_dgelist}
\alias{cqn_dgelist}
\title{cqn_dgelist}
\usage{
cqn_dgelist(x, length_column = "length", gc_column = "gc_percent",
  lib_size_column = "lib.size", ...)
}
\arguments{
\item{x}{A DGEList object}

\item{length_column}{Which column of x$genes contains the
gene-lengths.}

\item{gc_column}{Which column of x$genes contains the GC content?}

\item{lib_size_column}{Which column of x$samples contains the
library sizes?}

\item{...}{Further arguments to pass along to `cqn::cqn`}
}
\description{
cqn_dgelist
}
