% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_cutadapt_io.R
\name{define_cutadapt_summary_renaming}
\alias{define_cutadapt_summary_renaming}
\title{Define the renaming of the fields present in a cutadapt log summary section}
\usage{
define_cutadapt_summary_renaming()
}
\description{
Define the renaming of the fields present in a cutadapt log summary section
}
