% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_test_summary-methods.R
\name{export.default}
\alias{export.default}
\title{Default function for exporting a data-structure.}
\usage{
\method{export}{default}(x, ...)
}
\description{
Default function for exporting a data-structure.
}
