% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_test_summary-methods.R
\name{export}
\alias{export}
\title{Export a data structure to a (set of) file(s)}
\usage{
export(x, ...)
}
\arguments{
\item{x}{Some datastructure. Typically a
`sig_test_summary` as created by `reeq::get_sig_test_summary`.}

\item{...}{Further arguments. Currently unused.}
}
\description{
Export a data structure to a (set of) file(s)
}
