% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_test_summary-methods.R
\name{export.sig_test_summary}
\alias{export.sig_test_summary}
\title{Export a `sig_test_summary` object to files}
\usage{
\method{export}{sig_test_summary}(x, output_dir, test_name, ...)
}
\arguments{
\item{x}{Some datastructure. Typically a
`sig_test_summary` as created by `reeq::get_sig_test_summary`.}

\item{output_dir}{Which directory should the results be output to}

\item{test_name}{A name for the hypothesis test that is
summarised in the dataset `x`.}

\item{...}{Further arguments. Currently unused.}
}
\description{
Write the (unfiltered) top-table and significant gene lists from a
`sig_test_summary` object to some files. The file format is
`<output_dir>/<test_name>.top_tags.tsv` or
`<output_dir>/<test_name>.sig_features.<p_threshold>.tsv`, where the p-value
threshold is obtained from inside the `sig_test_summary` object and is
formatted to 4 decimal places.
}
