% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_cutadapt_io.R
\name{extract_cutadapt_summary}
\alias{extract_cutadapt_summary}
\title{Extract the 'summary' section from the text from a cutadapt log-file.}
\usage{
extract_cutadapt_summary(x)
}
\arguments{
\item{x}{The raw text from a cutadapt logfile. This fails
if more than one text entry is present, so the user should not split the
file on newlines or anything like that before calling this function.}
}
\description{
Returns a single string.
}
\details{
Not exported.
}
