% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_test_summary-methods.R
\name{get_features}
\alias{get_features}
\title{Extract the features (genes) from an object}
\usage{
get_features(x, ...)
}
\arguments{
\item{x}{Some object. Typically a sig_test_summary}

\item{...}{Further arguments. Currently unused.}
}
\description{
Extract the features (genes) from an object
}
