% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_test_summary-methods.R
\name{get_features.sig_test_summary}
\alias{get_features.sig_test_summary}
\title{Extract the features from a sig_test_summary}
\usage{
\method{get_features}{sig_test_summary}(x, ...)
}
\arguments{
\item{x}{Some object. Typically a sig_test_summary}

\item{...}{Further arguments. Currently unused.}
}
\description{
Extract the features from a sig_test_summary
}
