% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_annotation.R
\name{get_gc_percent}
\alias{get_gc_percent}
\title{Obtain the GC percentage for a set of genes from a biomaRt database}
\usage{
get_gc_percent(feature_ids, mart, feature_column = "ensembl_gene_id",
  gc_column = "percentage_gc_content")
}
\arguments{
\item{feature_ids}{A set of gene IDs.}

\item{mart}{A biomaRt `Mart` Dataset object.}

\item{feature_column}{The column of the `mart` dataset that
contains IDs of the same type as the `feature_ids`.}

\item{gc_column}{The column of the `mart` dataset that contains
the GC percentage values for the genes / features.}
}
\description{
Obtain the GC percentage for a set of genes from a biomaRt database
}
