% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{check_validity_k_covered}
\alias{check_validity_k_covered}
\title{Checks the validity of input to k-coverage functions}
\usage{
check_validity_k_covered(mat, k)
}
\arguments{
\item{mat}{Some numeric matrix or all-numeric data.frame}

\item{k}{A single number}
}
\description{
Helper function - it just checks that a numeric `matrix` or `data.frame` is
  passed in and that a number is passed in; ensures validity of arguments to
  the `k`-coverage functions
}
