% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{runner_k_covered}
\alias{runner_k_covered}
\title{runner_k_covered}
\usage{
runner_k_covered(mat, k, lambda)
}
\arguments{
\item{mat}{Either a matrix, data.frame, DGEList or
ExpressionSet.}

\item{k}{A single number.}

\item{lambda}{Some function lambda(mat, k).}
}
\description{
Runs a function over a matrix/data.frame/DGEList/ExpressionSet. Returns a
subset of the row indices for the matrix-like datastructure.
}
