% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_model.R
\name{import_contrasts}
\alias{import_contrasts}
\title{Import a contrasts matrix from a tab-separated file}
\usage{
import_contrasts(filepath, design)
}
\arguments{
\item{filepath}{a tab-separated file that defines the contrasts
for the current experiment.}

\item{design}{a design matrix. All design coefficients that
are used in defining contrasts should be column-names in this matrix.}
}
\description{
The `.tsv` file should contain two columns `contrast_name` and
`contrast_definition`. Contrast definitions should be in string form, eg,
"(group2 - group1)/2", where each word in the string should correspond to a
column in the `design` matrix.
}
\details{
Comment lines may begin with "#" in the .tsv
}
