% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_cutadapt_io.R
\name{import_cutadapt_summary}
\alias{import_cutadapt_summary}
\title{Extract the basepair/readpair counts from the 'summary' section in the text
from a cutadapt log-file.}
\usage{
import_cutadapt_summary(x)
}
\arguments{
\item{x}{A filename. This should be a cutadapt logfile.
The function dies if this is not an existing file or if more than one
entry is present in \code{x}.}
}
\description{
Extract the basepair/readpair counts from the 'summary' section in the text
from a cutadapt log-file.
}
