% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_hisat2_io.R
\name{import_hisat2_summary}
\alias{import_hisat2_summary}
\title{Extract the alignment summary statistics from a hisat2 logfile}
\usage{
import_hisat2_summary(x)
}
\arguments{
\item{x}{A filepath. This should be a hisat2 logfile. The
function dies if this is not an existing file of if more than one entry is
present in \code{x}.}
}
\description{
Extract the alignment summary statistics from a hisat2 logfile
}
