% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_cutadapt_io.R
\name{parse_cutadapt_summary}
\alias{parse_cutadapt_summary}
\title{Takes the 'summary' text from a cutadapt log-file and extracts the number of
readpairs / basepairs that were input and which pass/fail various filters.}
\usage{
parse_cutadapt_summary(x)
}
\arguments{
\item{x}{Text from the summary section of a cutadapt log.}
}
\description{
Not exported
}
