% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_hisat2_io.R
\name{parse_hisat2_summary}
\alias{parse_hisat2_summary}
\title{parse_hisat2_summary}
\usage{
parse_hisat2_summary(x)
}
\arguments{
\item{x}{The text from a hisat2 summary file. It is
assumed that the `--new-summary` flag was used while making this file.
There should be only a single entry in `x`.}
}
\description{
parse_hisat2_summary
}
