% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_helpers.R
\name{parse_numeric_fields}
\alias{parse_numeric_fields}
\title{parse_numeric_fields}
\usage{
parse_numeric_fields(x)
}
\arguments{
\item{x}{A single character string. This should be
newline-separated. Key-Value pairs are assumed to be colon-separated
and any line that is, is converted into a key-value pair.}
}
\value{
A dataframe. Contains columns "field" and "value" where
  the entries of \code{value} are the values found within the text \code{x}
  for the fields in \code{field}. The fieldnames _may_ be duplicated but are
  returned in the same order as they are observed in the logfile.
}
\description{
parse_numeric_fields
}
