% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_counts_io.R
\name{read_feature_counts}
\alias{read_feature_counts}
\title{Import read counts from a set of featureCounts files and combine them into
a single `data.frame`}
\usage{
read_feature_counts(files, sample_ids, comment = "#", col_types = "cii",
  progress = FALSE)
}
\arguments{
\item{files}{A vector of file-paths. Each should be an output
from `featureCounts` and contain, at least, the `Geneid`, `Length` and
feature-count columns.}

\item{sample_ids}{A vector of sample IDs corresponding, in order,
to the entries in `files`. If `NULL` ...}

\item{comment, col_types, progress}{Arguments passed through to
`readr::read_tsv` that specify the format of the input files and whether
a progress bar should be shown.}
}
\description{
Import read counts from a set of featureCounts files and combine them into
a single `data.frame`
}
