% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_counts_io.R
\name{read_single_feature_counts_file}
\alias{read_single_feature_counts_file}
\title{Read a single featureCounts file}
\usage{
read_single_feature_counts_file(file, comment = "#", col_types = "cii",
  progress = FALSE)
}
\arguments{
\item{file}{A single file, containing the `Geneid`, `Length`
and count columns from a single run of `featureCounts`.}

\item{comment}{Arguments passed through to
`readr::read_tsv` that specify the format of the input files and whether
a progress bar should be shown.}

\item{col_types}{Arguments passed through to
`readr::read_tsv` that specify the format of the input files and whether
a progress bar should be shown.}

\item{progress}{Arguments passed through to
`readr::read_tsv` that specify the format of the input files and whether
a progress bar should be shown.}
}
\description{
Read a single featureCounts file
}
