% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgelist_functions.R
\name{set_offset}
\alias{set_offset}
\title{Add (or replace) an offset matrix in DGEList}
\usage{
set_offset(dge, offset)
}
\arguments{
\item{dge}{A DGEList.}

\item{offset}{A matrix of the same dimensions as the counts in
`dge`.}
}
\value{
A DGEList.
}
\description{
Trivial implementation, but useful in pipelines.
}
