% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_cutadapt_io.R
\name{spread_and_rename_cutadapt_fieldnames}
\alias{spread_and_rename_cutadapt_fieldnames}
\title{spread_and_rename_cutadapt_fieldnames}
\usage{
spread_and_rename_cutadapt_fieldnames(x)
}
\arguments{
\item{x}{A dataframe with columns "field" and "value".}
}
\description{
This should be a temporary function
}
