% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_hisat2_io.R
\name{spread_and_rename_hisat2_fieldnames}
\alias{spread_and_rename_hisat2_fieldnames}
\title{spread_and_rename_hisat2_fieldnames}
\usage{
spread_and_rename_hisat2_fieldnames(x)
}
\arguments{
\item{x}{A datafame with two columns "field" and "value".
The numeric fields from a hisat2 logfile should have been extracted into
this dataframe.}
}
\description{
spread_and_rename_hisat2_fieldnames
}
