% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{which_k_covered}
\alias{which_k_covered}
\title{Returns the row-index for those rows in a `matrix` (or related: `DGEList` /
`ExpressionSet`) where at least 100`f` % of the columns have a value `>= k`.
Here, `f` is the argument `fraction_of_samples`.}
\usage{
which_k_covered(mat, k, fraction_of_samples = 1)
}
\arguments{
\item{mat}{Either a matrix, data.frame, DGEList or
ExpressionSet.}

\item{k}{A single number.}

\item{fraction_of_samples}{What fraction of the entries in a row
of `mat` should have a value `>= k` for that row to be included in the
results?}
}
\description{
Returns the row-index for those rows in a `matrix` (or related: `DGEList` /
`ExpressionSet`) where at least 100`f` % of the columns have a value `>= k`.
Here, `f` is the argument `fraction_of_samples`.
}
