% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{which_k_covered_across_samples}
\alias{which_k_covered_across_samples}
\title{which_k_covered_across_samples}
\usage{
which_k_covered_across_samples(mat, k = 2 * ncol(mat))
}
\arguments{
\item{mat}{Either a matrix, data.frame, DGEList or
ExpressionSet.}

\item{k}{A single number.}
}
\description{
Returns the row indices for those rows in a matrix where the sum of the
entries in that row is `>= k`. For example, in RNA-Seq, this returns the
index of those features that have a coverage `>= k` when summed across all
samples.
}
