% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgelist_functions.R
\name{append_feature_annotations}
\alias{append_feature_annotations}
\title{Add new columns from a table of feature-data into the feature-data
(`$genes`) of a `DGEList`}
\usage{
append_feature_annotations(dge, annotations, feature_id = "feature_id")
}
\arguments{
\item{dge}{A DGEList}

\item{annotations}{A data-frame to either use as the `genes` entry
or to append to the `genes` entry of the DGEList. Will match its
`feature_id` column against the feature ordering in the DGEList.}

\item{feature_id}{Which column of `annotations` contains the
IDs for the features. The contents of this column will be matched against
a column of `dge$genes` - this is preferentially a column of the same name
but, if that colname is absent from `dge$genes`, then the rownames of the
DGEList will be used.}
}
\value{
a DGEList
}
\description{
If the `DGEList$genes` entry is undefined, the feature data is added as the
genes entry; otherwise the new columns are appended to the RHS of the
existing genes entry.
}
\details{
Ensures that the feature-id / gene-id matches up between the provided
data-frame and the existing `DGEList`

The column named by `feature_id` must be present in the `annotations` data
frame. If that column is present in the $genes entry of `dge`, matching will
be done on that column, otherwise, matching will be done on the rownames of
the DGEList.
}
