% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_counts.R
\name{cbind_feature_counts}
\alias{cbind_feature_counts}
\title{cbind_feature_counts}
\usage{
cbind_feature_counts(count_list, names_as_colnames = FALSE)
}
\arguments{
\item{count_list}{List of data.frames. Each data.frame should have
columns "feature_id", "Length" and a final column that contains the read
counts for some sample.}

\item{names_as_colnames}{Boolean. If TRUE, the column ids in the
output dataframe are taken from the list names in `count_list`.}
}
\description{
Combines a list of data.frames into a single data.frame. The input
data.frames are of the format generated by featureCounts: the columns are
feature_id, Length and a file.path. The file.path column contains the read count
for a given feature in that sample. Returns a data.frame with the count
column for each sample juxtaposed and with feature_id and Length columns.
}
