% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_contrast_set_io.R
\name{parse_lrt_table}
\alias{parse_lrt_table}
\title{parse_lrt_table}
\usage{
parse_lrt_table(x)
}
\arguments{
\item{x}{A data-frame containing (at least) columns "lrt_name", "level",
"parent", "contrast_set". The `contrast_set` column must be a vector of
';'-separated characters or a list of character-vectors.}
}
\description{
Each likelihood-ratio test is performed over a set of contrasts. This
function extracts the names of the contrasts over which each LRT should be
performed from a data.frame. The collection of contrasts should be defined
as a ';'-separated string (or list of character-vectors) in the input
data.frame.
}
\details{
A tibble where the contrast sets are present in a list-column is returned.

The number of siblings of each LRT-node is also determined.

The `level` should be 1 for the LRT-nodes comprised of the largest
collection of contrasts and then increments for each proper subset of these
nodes. level-1 nodes have no parent, so set parent = NA for these rows.

The tibble that is returned is ordered by increasing 'level', to try to
ensure that parent nodes occur before children.

We allow a node to have at most one parent. For use in nested-LRT analysis,
any node that has a missing (NA) parent must be at level-1.
}
